pyproject.toml
[tool.poetry]
name = "articles-api"
version = "0.1.0"
description = "The application is based on two servers. API server provides information about users and their articles."
authors = ["Dddarknight <9801677@gmail.com>"]
license = "GPL-3.0-only"
keywords = ["fastAPI"]
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3",
  "Framework :: FastAPI"
]
repository = "https://github.com/Dddarknight/articles-api"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
SQLAlchemy = "^1.4.41"
databases = {extras = ["postgresql"], version = "^0.6.1"}
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
asyncpg = "^0.26.0"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^4.0.0"
alembic = "^1.8.1"
pymongo = "^4.2.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
